Updates after adding InlineHTMLPlugin
Move removeShebang to Parser, it's likely to be used more commonly with that class

Update formatting
