+++ src/Phan/AST/Parser.php
@@ -238,6 +238,39 @@ class Parser
         return $node;
     }
 
+    /**
+     * Remove the leading #!/path/to/interpreter/of/php from a CLI script, if any was found.
+     */
+    public static function removeShebang(string $file_contents) : string
+    {
+        if (substr($file_contents, 0, 2) !== "#!") {
+            return $file_contents;
+        }
+        for ($i = 2; $i < strlen($file_contents); $i++) {
+            $c = $file_contents[$i];
+            if ($c === "\r") {
+                if (($file_contents[$i + 1] ?? '') === "\n") {
+                    $i++;
+                    break;
+                }
+            } elseif ($c === "\n") {
+                break;
+            }
+        }
+        if ($i >= strlen($file_contents)) {
+            return '';
+        }
+        $rest = (string)substr($file_contents, $i + 1);
+        if (strcasecmp(substr($rest, 0, 5), "<?php") === 0) {
+            // declare(strict_types=1) must be the first part of the script.
+            // Even empty php tags aren't allowed prior to it, so avoid adding empty tags if possible.
+            return "<?php\n" . substr($rest, 5);
+        }
+        // Preserve the line numbers by adding a no-op newline instead of the removed shebang
+        return "<?php\n?>" . $rest;
+    }
+
+
     private static function shouldUsePolyfill(string $file_path, Request $request = null) : bool
     {
         if (Config::getValue('use_polyfill_parser')) {
